requirements.txt
flask==3.1.3
flask-limiter==4.1.1
groq==1.1.1
requests==2.31.0
python-dotenv==1.0.0